create and import fetchEvent
