+++ config/dbConfig.js
@@ -1,21 +1,21 @@
 const mysql = require("mysql2");
 require("dotenv").config();
 
-const pool = mysql.createPool({
-  host: process.env.DB_HOST,
-  user: process.env.DB_USER,
-  password: process.env.DB_PASSWORD,
-  database: process.env.DB_NAME,
-  port: 3306,
-});
-
 // const pool = mysql.createPool({
-//   host: "localhost",
-//   user: "root",
-//   password: "root",
-//   database: "ticket94",
+//   host: process.env.DB_HOST,
+//   user: process.env.DB_USER,
+//   password: process.env.DB_PASSWORD,
+//   database: process.env.DB_NAME,
 //   port: 3306,
 // });
+
+const pool = mysql.createPool({
+  host: "localhost",
+  user: "root",
+  password: "root",
+  database: "ticket94",
+  port: 3306,
+});
 
 const promisePool = pool.promise();
