+++ static/area.html
@@ -114,6 +114,7 @@
   </footer>
   <div id="toast" class="toast"></div>
   <script type="module" src="/scripts/area.js"></script>
+  <script type="module" src="/scripts/fetchEvent.js"></script>
   <script type="module" src="/scripts/signin-signup.js" defer></script>
 </body>
 </html>
