+++ static/booking.html
@@ -126,6 +126,7 @@
   </footer>
   <div id="toast" class="toast"></div>
   <script type="module" src="/scripts/booking.js"></script>
+  <script type="module" src="/scripts/fetchEvent.js"></script>
   <script type="module" src="/scripts/signin-signup.js" defer></script>
 
 </body>
