+++ static/checkout.html
@@ -120,6 +120,7 @@
   </footer>
   <div id="toast" class="toast"></div>
   <script type="module" src="/scripts/checkout.js"></script>
+  <script type="module" src="/scripts/fetchEvent.js"></script>
   <script type="module" src="/scripts/signin-signup.js" defer></script>
 </body>
 </html>
