+++ static/finish.html
@@ -72,6 +72,7 @@
     <div id="toast" class="toast"></div>
 
     <script type="module" src="/scripts/finish.js"></script>
+    <script type="module" src="/scripts/fetchEvent.js"></script>
     <script type="module" src="/scripts/signin-signup.js" defer></script>
 
   </body>
